backend/requirements.txt
gunicorn
asgiref==3.10.0
dj-database-url==3.0.1
Django==5.2.8
django-cors-headers==4.9.0
psycopg2-binary==2.9.11
sqlparse==0.5.3
tzdata==2025.2
whitenoise==6.11.0
djangorestframework==3.16.1
pillow==11.3.0
boto3==1.40.74
django-storages==1.14.6